requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.1
SQLAlchemy==2.0.35
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
python-dotenv==1.0.1
gunicorn==21.2.0
